+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Refactor image card CSS ([PR #5147](https://github.com/alphagov/govuk_publishing_components/pull/5147))
 * Decrease inverse header mobile/tablet padding ([PR #5149](https://github.com/alphagov/govuk_publishing_components/pull/5149))
 * Add option to inverse header for HMRC manuals ([PR #5148](https://github.com/alphagov/govuk_publishing_components/pull/5148))
+* **BREAKING** Remove youtube functionality from image card ([PR #5146](https://github.com/alphagov/govuk_publishing_components/pull/5146))
 
 ## 62.1.0